optimization (chunk) : uses shorts instead of floats for positions
This reduces the size the the buffers OpenGL buffer which help uploading them faster.
The side effect being that now ths position of the blocks is limited between (-2^15) and (2^15 - 1).
Which is big enough... But workarounds can be found by always rendering around 0...

+++ src/main/resources/shaders/chunk.vert
@@ -1,6 +1,6 @@
 #version 330
 
-layout(location = 0) in vec4 vPosition;
+layout(location = 0) in ivec4 vPosition;
 layout(location = 1) in vec2 vTexCoords;
 
 out vec2 passTexCoords;
